requirements.txt
streamlit
pandas
plotly
python-dotenv
browser-use
rapidfuzz
millify